feat(workspace): implement plugin engine, legal compliance, and sync
This release introduces several major architectural components and enterprise-ready features:

- **WebAssembly Extensibility**: Added `marksmen-plugin` providing a sandboxed Wasm engine via `wasmtime` for custom AST transformations.
- **Legal Compliance**: Implemented destructive redaction (`<redact>`) in Typst and DOCX pipelines to ensure sensitive data is physically stripped from output.
- **Cloud Synchronization**: Finalized `marksmen-cite` synchronization with a Last-Writer-Wins (LWW) merge engine for remote `references.json` management.
- **Interactive Form Synthesis**: Added support for bridging HTML `<form>` elements into native Word Content Controls and PDF markers.
- **Browser Ingestion**: Deployed a new browser extension for scraping citation metadata directly into the local `marksmen-cite` server.
- **Security Enhancements**: Added support for PDF encryption and digital signatures via `marksmen-crypto`.
- **Core Improvements**:
    - Optimized memory allocation in `AstConcatenator` to prevent fragmentation.
    - Added XObject recursion depth guards to the native PDF reader to prevent stack overflows.
    - Improved Mermaid rendering in ODT via SVG-to-PNG conversion.
    - Refactored several translation engines to use more efficient iterator patterns.

Updated workspace configuration to exclude Tauri-specific crates from standard `cargo test` runs.

+++ crates/marksmen-pdf-read/src/reader/font/metrics.rs
@@ -185,7 +185,7 @@ impl WidthTable {
 
 fn obj_to_f32(o: &Object) -> f32 {
     match o {
-        Object::Real(f) => *f as f32,
+        Object::Real(f) => *f,
         Object::Integer(i) => *i as f32,
         _ => 0.0,
     }
@@ -209,7 +209,7 @@ fn descriptor_missing_width(font_dict: &Dictionary, doc: &Document) -> Option<f3
     };
     fd.get(b"MissingWidth").ok().and_then(|o| match o {
         Object::Integer(i) => Some(*i as f32),
-        Object::Real(f) => Some(*f as f32),
+        Object::Real(f) => Some(*f),
         _ => None,
     })
 }
